extract stats collector
