+++ packages/ios-client/Sources/FishjamClient/webrtc/PeerConnectionManager.swift
@@ -16,8 +16,13 @@ internal class PeerConnectionManager: NSObject, RTCPeerConnectionDelegate {
         }
     }
 
-    private var connection: RTCPeerConnection?
-    private var peerConnectionStats: [String: RTCStats] = [:]
+    private var statsCollector: StatsCollector?
+    private var connection: RTCPeerConnection? {
+        didSet {
+            guard let connection = connection else { return }
+            statsCollector = StatsCollector(connection: connection)
+        }
+    }
 
     private var iceServers: [RTCIceServer] = []
     private var config: RTCConfiguration?
@@ -197,7 +202,7 @@ internal class PeerConnectionManager: NSObject, RTCPeerConnectionDelegate {
         if let pc = connection {
             pc.close()
             connection = nil
-            peerConnectionStats = [:]
+            statsCollector = nil
             iceServers = []
             config = nil
             midToTrackId = [:]
@@ -304,55 +309,9 @@ internal class PeerConnectionManager: NSObject, RTCPeerConnectionDelegate {
         sender.parameters = params
     }
 
-    private func extractRelevantStats(rp: RTCStatisticsReport) {
-        rp.statistics.forEach { it1 in
-            let it = it1.value
-            if it.type == "outbound-rtp" {
-                let duration = it.values["qualityLimitationDurations"] as? [String: Double]
-                let qualityLimitation: QualityLimitationDurations = QualityLimitationDurations(
-                    bandwidth: duration?["bandwidth"] ?? 0.0,
-                    cpu: duration?["cpu"] ?? 0.0, none: duration?["none"] ?? 0.0, other: duration?["other"] ?? 0.0)
-
-                let tmp = RTCOutboundStats(
-                    kind: it.values["kind"] as? String ?? "",
-                    rid: it.values["rid"] as? String ?? "",
-                    bytesSent: it.values["bytesSent"] as? UInt ?? 0,
-                    targetBitrate: it.values["targetBitrate"] as? Double ?? 0.0,
-                    packetsSent: it.values["packetsSent"] as? UInt ?? 0,
-                    framesEncoded: it.values["framesEncoded"] as? UInt ?? 0,
-                    framesPerSecond: it.values["framesPerSecond"] as? Double ?? 0.0,
-                    frameWidth: it.values["frameWidth"] as? UInt ?? 0,
-                    frameHeight: it.values["frameHeight"] as? UInt ?? 0,
-                    qualityLimitationDurations: qualityLimitation
-                )
-
-                peerConnectionStats[it.id as String] = tmp
-            } else if it.type == "inbound-rtp" {
-                let tmp = RTCInboundStats(
-                    kind: it.values["kind"] as? String ?? "",
-                    jitter: it.values["jitter"] as? Double ?? 0.0,
-                    packetsLost: it.values["packetsLost"] as? UInt ?? 0,
-                    packetsReceived: it.values["packetsReceived"] as? UInt ?? 0,
-                    bytesReceived: it.values["bytesReceived"] as? UInt ?? 0,
-                    framesReceived: it.values["framesReceived"] as? UInt ?? 0,
-                    frameWidth: it.values["frameWidth"] as? UInt ?? 0,
-                    frameHeight: it.values["frameHeight"] as? UInt ?? 0,
-                    framesPerSecond: it.values["framesPerSecond"] as? Double ?? 0.0,
-                    framesDropped: it.values["framesDropped"] as? UInt ?? 0
-                )
-
-                peerConnectionStats[it.id as String] = tmp
-            }
-        }
-    }
 
     public func getStats() -> [String: RTCStats] {
-        if let connection = connection {
-            connection.statistics(completionHandler: { RTCStatisticsReport in
-                self.extractRelevantStats(rp: RTCStatisticsReport)
-            })
-        }
-        return peerConnectionStats
+        return statsCollector?.peerConnectionStats ?? [:]
     }
 
     public func getSdpOffer(
